[update] 网盟实体更新

+++ perfect-entity/src/main/java/com/perfect/entity/report/NmsAccountReportEntity.java
@@ -47,7 +47,7 @@ public class NmsAccountReportEntity extends AccountIdEntity {
     private Integer srchuv;     // 展现独立访客
 
     @Field(value = "clickuv")
-    private Integer clickuv;    // 点击独立访客
+    private Double clickuv;    // 点击独立访客
 
     @Field(value = "srsur")
     private Integer srsur;  // 展现频次
@@ -153,11 +153,11 @@ public class NmsAccountReportEntity extends AccountIdEntity {
         this.srchuv = srchuv;
     }
 
-    public Integer getClickuv() {
+    public Double getClickuv() {
         return clickuv;
     }
 
-    public void setClickuv(Integer clickuv) {
+    public void setClickuv(Double clickuv) {
         this.clickuv = clickuv;
     }
