+++ perfect-entity/src/main/java/com/perfect/entity/report/NmsGroupReportEntity.java
@@ -52,7 +52,7 @@ public class NmsGroupReportEntity extends AccountIdEntity {
     private Integer clickuv;    // 点击独立访客
 
     @Field(value = "srsur")
-    private Integer srsur;  // 展现频次
+    private Double srsur;  // 展现频次
 
     @Field(value = "cusur")
     private Double cusur;  // 独立访客点击率
@@ -179,11 +179,11 @@ public class NmsGroupReportEntity extends AccountIdEntity {
         this.clickuv = clickuv;
     }
 
-    public Integer getSrsur() {
+    public Double getSrsur() {
         return srsur;
     }
 
-    public void setSrsur(Integer srsur) {
+    public void setSrsur(Double srsur) {
         this.srsur = srsur;
     }
